Reorganizacion de los archivos en las carpetas correspondientes
