+++ .gitignore
@@ -1,3 +1,11 @@
-# Ignora carpetas y sus archivos
+# Default ignored files
+/shelf/
+/.idea/workspace.xml
+# Editor-based HTTP Client requests
+/httpRequests/
+# Datasource local storage ignored files
+/dataSources/
+/dataSources.local.xml
 .idea/*
 allure-results/*
+target/*
