--- src/main/java/DemoTest/PruebaSelenium.java
@@ -1,5 +1,0 @@
-package DemoTest;
-
-public class PruebaSelenium {
-
-}
